Day 6, extract kudocracy code
To get back to were I was in early 2018, I extracted some files from project Kudocracy and provided a compatibility layer in functions/kudocracy.js

This works up to the point where external package twit is invoked. That package is no longer maintaned, I need to swith to the younger twitter-lite package.